refactor: move rendering class to core
