+++ src/demo/index.js
@@ -1,6 +1,6 @@
 import "./style.css"
 import { gsap } from "gsap"
-import { Rendering } from "./rendering.js"
+import { Rendering } from "../core/rendering.js"
 import * as THREE from "three";
 
 import { OrbitControls } from "three/examples/jsm/controls/OrbitControls.js"
